+++ src/vue/modules/movements-history/components/operation-viewers/payment.vue
@@ -57,7 +57,9 @@
       <td class="attributes-viewer__table-cell">
         {{ 'movements-history.subject-lbl' | globalize }}
       </td>
-      <td class="attributes-viewer__table-cell subject-table-cell">
+      <td
+        class="attributes-viewer__table-cell
+          attributes-viewer__subject-table-cell">
         {{ operationDetails.subject }}
       </td>
     </tr>
@@ -160,7 +162,7 @@ export default {
 <style lang="scss" scoped>
   @import '../../scss/attributes-viewer';
 
-  .subject-table-cell {
+  .attributes-viewer__subject-table-cell {
     word-break: break-word;
     white-space: normal;
   }